requirements.txt
streamlit
torch
torchvision
torchmetrics
timm
Pillow
numpy
matplotlib
pyyaml
pytorch_lightning
ultralytics
streamlit_cropper
opencv-python-headless
pytorch-grad-cam
ultralytics
cv2
shapely
gdown